# app/openapi-resolved.yaml
openapi: 3.0.3
info:
  title: 選課清單
  version: 1.0.0
servers:
- url: https://irisyen115.synology.me/course-api
paths:
  /courses:
    get:
      summary: 取得課程列表（含講師基本資訊）
      responses:
        "200":
          description: 成功取得課程資料
          content:
            application/json:
              schema:
                maxItems: 2
                type: array
                items:
                  $ref: "#/components/schemas/CourseWithInstructor"
  /course:
    post:
      summary: 建立新課程
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
        required: true
      responses:
        "201":
          description: 課程建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /course/{courseId}:
    put:
      summary: 更新課程內容
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
        required: true
      responses:
        "200":
          description: 課程更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      summary: 刪除課程
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: 課程刪除成功（無內容回應）
  /instructors:
    get:
      summary: 取得授課講師列表
      responses:
        "200":
          description: 成功取得講師列表
          content:
            application/json:
              schema:
                maxItems: 2
                type: array
                items:
                  $ref: "#/components/schemas/Instructor"
  /instructor:
    post:
      summary: 建立新講師
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstructorInput"
        required: true
      responses:
        "201":
          description: 講師建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
  /instructor/{instructorId}/courses:
    get:
      summary: 取得指定講師開設的課程（最多兩筆）
      parameters:
      - name: instructorId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功取得講師的課程
          content:
            application/json:
              schema:
                maxItems: 2
                type: array
                items:
                  $ref: "#/components/schemas/Course"
components:
  schemas:
    Instructor:
      type: object
      properties:
        id:
          type: string
          example: instr_001
        name:
          type: string
          example: 王小明
        email:
          type: string
          format: email
          example: ming@example.com
    InstructorInput:
      required:
      - email
      - name
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    Course:
      type: object
      properties:
        id:
          type: string
          example: course_001
        title:
          type: string
          example: Python 入門班
        description:
          type: string
          example: 從零開始學 Python
        instructorId:
          type: string
          example: instr_001
    CourseInput:
      required:
      - description
      - instructorId
      - title
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        instructorId:
          type: string
    CourseWithInstructor:
      allOf:
      - $ref: "#/components/schemas/Course"
      - type: object
        properties:
          instructor:
            $ref: "#/components/schemas/Instructor"